backend/requirements.txt
asgiref==3.8.1
authentication==1.1.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
redis==5.0.8
sqlparse==0.5.1
tzdata==2024.1
Werkzeug==3.0.3
WTForms==3.1.2